emotion-detection/backend/requirements.txt
flask>=3.0,<4.0
flask-cors>=5.0,<7.0
torch>=1.10,<3.0
numpy>=2.1,<3.0
scipy>=1.14,<2.0
librosa>=0.11,<1.0
soundfile>=0.12,<1.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0

